Match implementation to API schema
